Remove the Rust solution because it's too awkward
Rust libraries are generally cross-platform, which makes some of the specific library dynamic loading etc. too difficult. For example, I need to get an HMODULE representing a DLL and pass it to the plugin.

And anyway, I'm working so closely with the plugin that there's a lot of C FFI interop, and it's just a lot of unnecessary work, and kinda hard to get it right too.