Create OutputPlugin interface and a subclass.
Each output plugin is a singleton, because the output plugin interface uses function pointers without context, so it's not possible to tell whether a particular function is being called for one instance or another. At least, different output plugin types can be distinguished (similar to how different output plugin DLLs can be distinguished), but this is unlikely to be useful to us.